refactor: remove last legacy code forms and formatting
