--- src/ConsoleScreenView.swift
@@ -1,20 +1,0 @@
-//
-//  ConsoleScreenView.swift
-//  Page of the main view that serves as the console
-//
-//  This file may be distributed under the terms of the GNU GPLv3 license.
-//
-
-import SwiftUI
-
-struct ConsoleScreenView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct ConsoleScreenView_Previews: PreviewProvider {
-    static var previews: some View {
-        ConsoleScreenView()
-    }
-}
