+++ src/PrintersView.swift
@@ -81,48 +81,3 @@ struct PrinterCardView: View {
         }
     }
 }
-
-struct EditSheetView: View {
-    @Binding var newPrinterData: PrinterConfig.ModifiedData
-    @Binding var isInEditMode: Bool
-    @Binding var isPresentingEditSheet: Bool
-    @Binding var printerBeingEdited: Int
-    @Binding var printers: [PrinterConfig]
-
-    var body: some View {
-        NavigationView {
-            PrinterEditView(data: $newPrinterData, isInEditMode: $isInEditMode, deleteCall: {
-                printers.remove(at: printerBeingEdited)
-                isPresentingEditSheet = false
-            })
-            .toolbar() {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("Cancel") {
-                        isPresentingEditSheet = false
-                        newPrinterData = PrinterConfig.ModifiedData()
-                    }
-                }
-                ToolbarItem(placement: .confirmationAction) {
-                    Button("Save") {
-                        if (newPrinterData.name.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty){
-                            
-                            return
-                        }
-                        
-                        isPresentingEditSheet = false
-                        if (isInEditMode){
-                            printers.remove(at: printerBeingEdited)
-                            printers.insert(PrinterConfig(data: newPrinterData), at: printerBeingEdited)
-                            isInEditMode = false
-                        } else {
-                            printers.append(PrinterConfig(data: newPrinterData))
-                            newPrinterData = PrinterConfig.ModifiedData()
-                        }
-                        
-                    }
-                }
-            }
-            .navigationTitle(isInEditMode ? "Edit Printer":"Add Printer")
-        }
-    }
-}
